4.2.4 HTML additions Input and Select(list)
Made a <form> and moved the button into it giving it the type="submit.

<div class="form-group">
      <input type="text" name="task-name" placeholder="Enter Task Name"/>
    </div>
Added a text input type to inside this <form> so we can remember names

<div class="form-group">
      <select name="task-type" class="select-dropdown">
        <option value="" disabled selected>Pick a task type</option>
        <option value="Print">Print</option>
        <option value="Web">Web</option>
        <option value="Mobile">Mobile</option>
      </select>
Added a drop down list inside the <form> that we can click to select 3 of the 4 options.  The top one is disabled.

The CSS was largely automatic seeing as how they provided it but WE DID add class="select-dropdown" to the <select> element to use their prebuilt CSS at the tail end