4.2.5
var buttonEl = document.querySelector("#save-task");
var formEl = document.querySelector("#task-form");
Changed the name of the El because we are targeting a new id #task-form so we can pull the entire form element when clicking the button.

var createTaskHandler = function(event) {
  event.preventDefault();
Stops the default function from the page from happening.  By default the page cleared the text we created

formEl.addEventListener("submit", createTaskHandler);
targets the newly created formEL and changed to from "click" to "submit" with the addEventListener which targets the type
<button class="btn" id="save-task" type="submit">Add Task</button>
in the HTML.  This is supposed to allow us to click and hit enter.  Didn't seem to work but that was the intention.

Useful Vid about input at the top.