4.2.6
Learned how to append items into others.

var createTaskHandler = function (event) {
  event.preventDefault();
passed and event that stops the screen from refreshing

  var taskNameInput = document.querySelector("input[name='task-name']").value;
  console.log(taskNameInput);
this creates a var that then goes in and grabs what we type in the html element <input type="text" name="task-name".....>

  var taskTypeInput = document.querySelector("select[name='task-type']").value;
  console.log(taskTypeInput);
grabs what we selected from the element <select name="task-type"....> and puts it in the var taskTypeInput
  // create list item
  var listItemEl = document.createElement("li");
  listItemEl.className = "task-item";
creates a new li without the help of HTML and sets it to var listItemel.  Then the next line styles it in the CSS with a class selector named .task-item

  // create div to hold task info and add to list item
  var taskInfoEl = document.createElement("div");
  taskInfoEl.className = "task-info";
creates a div without the help of HTML to hold the info going into the li named taskInfoEl and the next line styles this div in the CSS with the class .task-info

  // add HTML content to div
  taskInfoEl.innerHTML = "<h3 class='task-name'>" + taskNameInput + "</h3><span class='task-type'>" + taskTypeInput + "</span>";
 this adds the text the divs var just created taskInfoEl using innerHTML instead of textContent.  Then each "<>" is simply calling CSS style for the the vars between the +var+.  textContent would ignore the CSS calls and show them instead as text instead of innerHTML which just translates it to style and just shows the vars text.

listItemEl.appendChild(taskInfoEl);
this adds(appends) the text we just created in this div to the li listItemEl.  This will not show yet seeing has li does not exsist in the HTML.  We have to......
  // add entire list item to list
  tasksToDoEl.appendChild(listItemEl);
};
this moves(appends) the entire var  listItemEl into var tasksToDoEl which has a home in the HTML thanks to the querySelector up top targeting the ul id #tasks-to-do and add it to the HTML as a result.
formEl.addEventListener("submit", createTaskHandler);