4.2.8
refactored code.  Major thing learned here was this.

// package up data as an object
  var taskDataObj = {
      name: taskNameInput,
      type: taskTypeInput
  };
we made an object in the top function taskFormHandler and passed it into the bottom function createTaskEl like so
createTaskEl(taskDataObj);
this is essentially a call with the object in the () and is giving it away and then the function below...
var createTaskEl = function (taskDataObj) {
receives the object by adding it to it's own ().

taskInfoEl.innerHTML = "<h3 class='task-name'>" + taskDataObj.name + "</h3><span class='task-type'>" + taskDataObj.type + "</span>";
this then created a moment needing me to change the vars for the innerHTML from  taskNameInput and taskTypeInput to taskDataObj.name and taskDataObj.type to be able to talk to the object being sent in as opposed to before when it had the vars within it's own functiony braces{}