4.2.9
if (!taskNameInput || !taskTypeInput) {
    alert("You need to fill out the task form!");
    return false;
  }
if statement that checks if the user selected nothing and yells at them if they haven't  in an alert.  Then it returns a false for reasons.

formEl.reset();
This resets the text in the text box after use to it's default state.  Only works on the form element.