4.3.5
var taskIdCounter = 0;
created a var set to zero for purposes of

 listItemEl.setAttribute("data-task-id", taskIdCounter);
assigning it to a list.  Basically every list we create with the will be assigned data-task-id set to 0 until we

taskIdCounter++;
below it add  this which will then go back up to the global scope and increase the value  of the var taskIdCounter by one.

Then the next time we press the button the value will be 1, then 2 etc..

Has a video that introduces the data attribute.