4.3.6
created the new function createTaskEl

var taskActionsEl = createTaskActions(taskIdCounter);
here we store the entire called function and it's number from taskIdCounter into the var taskActionsEl
  listItemEl.appendChild(taskActionsEl);
 then appended the new var taskActionsEl into the created list listItemEl   (just  after this it appends into the ul which makes it appear and incrments by one so the next list can have a new id)

var createTaskActions = function(taskId) {
this starts the function and changes the name of taskIdCounter to taskId for this function

  var actionContainerEl = document.createElement("div");
  actionContainerEl.className = "task-actions";
this creates and styles the box that everything is going to go in

  //create edit button
  var editButtonEl = document.createElement("button");
  editButtonEl.textContent = "Edit";
  editButtonEl.className = "btn edit-btn";
  editButtonEl.setAttribute("data-task-id", taskId);
this creates, adds text, styles and assigns the passed in id number to data-task-id
  actionContainerEl.appendChild(editButtonEl);
this adds it to the actionContainerEl div

  //create delete button
  var deleteButtonEl = document.createElement("button");
  deleteButtonEl.textContent = "Delete";
  deleteButtonEl.className = "btn delete-btn";
  deleteButtonEl.setAttribute("data-task-id", taskId);
this creates, gives text, styles and assigns the passed in number to data-task-id

  actionContainerEl.appendChild(deleteButtonEl);
this adds it to the actioContainerEl div

  //create dropdown
  var statusSelectEl = document.createElement("select");
  statusSelectEl.className = "select-status";
  statusSelectEl.setAttribute("name", "status-change");
  statusSelectEl.setAttribute("data-task-id", taskId);
this creates, names, sets the name to status change and assigns the passed number to data-task-id

  var statusChoices = ["To Do", "In Progress", "Completed"];
makes an array of a length of 3

  for (var i = 0; i < statusChoices.length; i++) {
checks which spot we are at in the array and then
    //create option element
    var statusOptionEl = document.createElement("option");
    statusOptionEl.textContent = statusChoices[i];
    statusOptionEl.setAttribute("value", statusChoices[i]);
creates an option element , assigns visible text using [i], assigns a number to this spot named value all based on it's location in the array
ie 0 would be To Do etc..

    //append select
    statusSelectEl.appendChild(statusOptionEl);
  }
adds the option elements (3 times in this case) to the select element to create a dropdown window.  then finishes it's for loop.

  actionContainerEl.appendChild(statusSelectEl);
adds the select element statusSelectElementEl to the main div of this function actionContainerEl

  return actionContainerEl;
then returns the completed div actionContainerEl to the above state that called it
};