4.3.7
var pageContentEl = document.querySelector("#page-content");
created a new a var pageContentEl that references the <main> element id

var taskButtonHandler = function(event) {
  console.log(event.target);
made a function that uses the event to allow the console.log to print the target ie what is clicked

  if(event.target.matches(".delete-btn")){
this then uses the word matches as well  to use this if statement to see if what is clicked is the button with a class of .delete-btn

    //get the element's task id
    var taskId = event.target.getAttribute("data-task-id");
    deleteTask(taskId);
  }
};
this then puts the data-task-id,  ie the current boxes id number, into the var taskId and then calls a new function deleteTask and passes var taskId into it.  Finally it closes the if statement and the function

var deleteTask = function(taskId) {
this is the start of the new function with the var the clicked on number passed in

  var taskSelected = document.querySelector(".task-item[data-task-id='" + taskId +"']");
this then creates a var that uses querySelector to select the created li's class and it's data-task-id which uses the var to pass in the the number of it's id.

  taskSelected.remove();
this simply removes the newly created var selected this specific li
};

pageContentEl.addEventListener("click", taskButtonHandler);
the call that references the <main> var in the html and calls taskButtonHandler using the click addEventListener

Video about Bubbling near the top