4.3.8
 var targetEl = event.target;
made a var for event.target because we are fancy.

if (targetEl.matches(".edit-btn")) {
    var taskId = targetEl.getAttribute("data-task-id");
    editTask(taskId);
  }
targets only the edit button using the new var and mataches then
puts the selected li's number from data-tak-id into the var taskId and then runs the edit function passing taskId to it

var editTask = function(taskId) {
start of edit function receiving the var from above

  //get task list item element
  var taskSelected = document.querySelector(".task-item[data-task-id='" + taskId +"']");
this then creates a var that uses querySelector to select the created li's class and it's data-task-id which uses the var to pass in the number of it's id.

  //get content from task name and type
  var taskName = taskSelected.querySelector("h3.task-name").textContent;
  var taskType = taskSelected.querySelector("span.task-type").textContent;
these put the text contents in their respective variables

  document.querySelector("input[name='task-name']").value = taskName;
  document.querySelector("select[name='task-type']").value = taskType;
puts the just created variable up top in the header showing the names that were in the <main>'s  selected  <li>

  document.querySelector("#save-task").textContent = "Save Task";
changes the text of "Add Task to "Save Task"
  formEl.setAttribute("data-task-id", taskId);
};
gives the number of the selected ID where this text came from to the <form> and closes the function

also added the new variable targetEl to the delete button instead of event.target just because we are fancy.

We used querySelector to select specific elements we created for the first time
ex..
var taskSelected = document.querySelector(".task-item[data-task-id='" + taskId +"']");
 var taskName = taskSelected.querySelector("h3.task-name").textContent;
 var taskType = taskSelected.querySelector("span.task-type").textContent;