4.3.9
Next three bits (var,if,else) are added to the taskFormHandler function

 var isEdit = formEl.hasAttribute("data-task-id");
this creates a variable to check if <form> is assigned a number from the data-task-id.  At the begining the answer will be no becaue we haven't gotten to the to the point where we assign one yet

  // has data attribute, so get task id and call function to complete edit process
if (isEdit) {
  var taskId = formEl.getAttribute("data-task-id");
  completeEditTask(taskNameInput, taskTypeInput, taskId);
}
this checks if the newly created var isEdit is true.
If there is a number it then it uses getAttribute to grab it's number from data-task-id and assigns to the var taskId.
Then it runs a newly created function to deal with this and passes in taskNameInput for what is typed there, taskTypeInput for what is selected and the just grabbed number in the var taskId.

// no data attribute, so create object as normal and pass to createTaskEl function
else {
  var taskDataObj = {
    name: taskNameInput,
    type: taskTypeInput
  };

  createTaskEl(taskDataObj);
This else statement means otherwise we function as normal which creates the  the object and runs the createTaskEl function passing in the newly created object.

When the above is statement is true then...
var completeEditTask = function(taskName, taskType, taskId) {
we run this newly created function and even change the names of the first to vars from taskNameInput to taskName and taskTypeInput to taskType and the taskId stays the same.

  //find the matching task list item
  var taskSelected = document.querySelector(".task-item[data-task-id='" + taskId + "']");
this finds a list using querySelector using the passed in taskID to find it's number and assigns it to var taskSelected

  //set new values
  taskSelected.querySelector("h3.task-name").textContent = taskName;
  taskSelected.querySelector("span.task-type").textContent = taskType;
we use the newly created var taskSelected to then use querySelector to find its element <h3> and its class  of .task-name.  Then we grab it's textContent and assign it to the newly renamed var taskName.  We then do the exact same thing for the next line with the <span> .task-type and put it's text-content into the newly renamed var taskType

  alert("Task Updated!");
an alert to let you know you updated the task

  formEl.removeAttribute("data-task-id");
removes the data-task-id number from the form using removeAttrbute so we can have the list create new li's as normal
  document.querySelector("#save-task").textContent = "Add Task";
changes the text-content of the button back to Add Task to appear is if it has reset back to normal.
}