4.3.10
var tasksInProgressEl = document.querySelector("#tasks-in-progress");
var tasksCompletedEl = document.querySelector("#tasks-completed");
added the ul's from the HTML to the JS

pageContentEl.addEventListener("change", taskStatusChangeHandler);
calls the new function by listening for a "change" which listens for when a <form> value changes

var taskStatusChangeHandler = function(event) {
start of new function passing in event so we can target things

  //get the task items id
  var taskId = event.target.getAttribute("data-task-id");
an example of looking at what we are targeting and then getting it's attribute which is the number we've assigned to data-task-id and bringing it into the func and then assigning it to the var taskId.

  //get the currently selected option's value and convert to lowercase
  var statusValue = event.target.value.toLowerCase();
gets the value of what is targeted in this case the name to do, tasks in progress or completed and sets it to lowercase to future proof it and assigns it to the var statusValue.

  //find the parent task item element based on the id
  var taskSelected = document.querySelector(".task-item[data-task-id='" + taskId + "']");
finds the li using querySelector and then gets specific with what list by taking the newly created var taskId and then assigns this entire specific list to the var taskSelected.

  if (statusValue === "to do") {
    tasksToDoEl.appendChild(taskSelected);
  }
this checks if the value of statusValue is "to do".  If so then it appends the entire li from it's current ul and moves it to the var tasksToDoEl  which references the id of the ul in HTML..

  else if (statusValue === "in progress") {
    tasksInProgressEl.appendChild(taskSelected);
  }
this checks if the value of statusValue is "in progress".  If so then it appends the entire li from it's current ul and moves it to the var  tasksInProgressEl which references the id of the ul in HTML..

  else if (statusValue === "completed") {
    tasksCompletedEl.appendChild(taskSelected);
  }
this checks if the value of statusValue is "completed".  If so then it appends the entire li from it's current ul and moves it to the var tasksCompletedEl  which references the id of the ul in HTML.

};
closes the func taskStatusChangeHandler

Big thing to note is when appending the li it does not create a copy.  JS needs document.createElement to do something like that.  It instead just finds the li we targeted and appends that existing element somewhere else.