4.4.4
var tasks = [];
added the new array to the global scope to the var tasks

 var taskDataObj = {
      name: taskNameInput,
      type: taskTypeInput,
      status: "to do"
    };
added status var to the object var taskDataObj

var createTaskEl = function(taskDataObj) {
  console.log(taskDataObj);
  console.log(taskDataObj.status);

  var listItemEl = document.createElement("li");
  listItemEl.className = "task-item";
  listItemEl.setAttribute("data-task-id", taskIdCounter);

  var taskInfoEl = document.createElement("div");
  taskInfoEl.className = "task-info";
  taskInfoEl.innerHTML = "<h3 class='task-name'>" + taskDataObj.name + "</h3><span class='task-type'>" + taskDataObj.type + "</span>";
  listItemEl.appendChild(taskInfoEl);

  // create task actions (buttons and select) for task
  var taskActionsEl = createTaskActions(taskIdCounter);
  listItemEl.appendChild(taskActionsEl);
  tasksToDoEl.appendChild(listItemEl);

  taskDataObj.id = taskIdCounter;
we assign the object taskDataObj an id using taskDataObj.id and put the passed in taskIdCounter number after the createTaskEl func and the called  createTaskAction func finish their work with it.  Then..

tasks.push(taskDataObj);
we take that info and use the tasks.push to push this newly formed taskDataObj  into the var tasks array  (push is used for arrays) with its name, type, status and id in tow.

var completeEditTask = function(taskName, taskType, taskId) {
  // find task list item with taskId value
  var taskSelected = document.querySelector(".task-item[data-task-id='" + taskId + "']");

  // set new values
  taskSelected.querySelector("h3.task-name").textContent = taskName;
  taskSelected.querySelector("span.task-type").textContent = taskType;

  //loop through tasks array and task object with new content
  for (var i = 0; i < tasks.length; i++) {
in the completeEditTask func  we step through the arrays length and
    if (tasks[i].id === parseInt(taskId)){
if the current number we are on matches the passed in taskId attribute.
Also we have to use parseInt to convert taskId to a number instead of a string.
      tasks[i].name = taskName;
      tasks[i].type = taskType;
we assign the new name and type to the array spot of [i] matching it's taskId replacing the previous Name and Type with the new text.
    }
  };

//update task's in tasks array
  for (var i = 0; i < tasks.length; i++) {
    if (tasks[i].id === parseInt(taskId)) {
      tasks[i].status = statusValue;
    }
  }
  console.log(tasks);
};
similar to before but in the taskStatusChangeHandler func we will make a for loop to the name of the statusValue var and assign it to the array using tasks[i].sataus to reference which which array change the status within it.

deleteTask func--
var updatedTaskArr = [];
creates an empty array to the var updatedTaskArr = [];

  //loop through current tasks
  for (var i = 0; i< tasks.length; i++) {
    //if tasks[i].id doesn't match the value of taskId, let's keep that task and push it into the new array
    if (tasks[i].id !== parseInt(taskId)) {
if the current spot the for loop is on is NOT the current taskId.  Then we go ahead and
      updatedTaskArr.push(tasks[i]);
add the current obj to this new array
    }
if statement done
  }
for loop done with one less

  //reassign tasks array to be the same as updatedTaskArr
  tasks = updatedTaskArr;
takes the new array var updatedTaskArray and pastes over the global task array