4.4.6
+var loadTasks = function() {
new func called loadTasks to get the array out of local storage by

  tasks = localStorage.getItem("tasks");
goes into the local storage with localStorage.getItem and grabs the tasks array as a string as it was stored with JSON.stringify

  // if thereare no tasks, set tasks to an empty array and return out of the function
   if (!tasks) {
    tasks=[];
    return false;
  }
this is to stop this function from breaking the tasks array from breaking.  By default if the array is empty this load would return a null which breaks other functions.  Here we change it to an empty array and leave the func with a value of false.

  tasks = JSON.parse(tasks);
 otherwise this will then take the grabbed string array and convert it to the value or object it was before the save func made it a string.

  for (i=0; i < tasks.length; i++) {
this walks through the newly loaded array one step at a time starting with 0  and contining on to

    var listItemEl = document.createElement("li");
    listItemEl.className = "task-item";
    listItemEl.setAttribute("data-task-id", tasks[i].id);
create the list and give it a class and it's data task id number.  The attribute would look something like data-task-id = 0 if in the first spot

    var taskInfoEl = document.createElement("div");
    taskInfoEl.className = "task-info";
    taskInfoEl.innerHTML = "<h3 class='task-name'>" + tasks[i].name + "</h3><span class='task-type'>" + tasks[i].type + "</span>";
this then creates the div for the name and type that are stored in the array, gives it a class of task info and sets the inner HTML to to show the text name in an h3 and the text from type in a span ele all within this div.

    listItemEl.appendChild(taskInfoEl);
this then adds this newly created <div> of taskInfoEl and adds it to the <li> listItemEl above.

    taskActionsEl = createTaskActions(tasks[i].id);
for this next bit we take a different approach.  Intead of creating both <button>s and the <select> when instead call the og func  that did that in the first place passing in the arrays id number.  We then take the edit, delete and drop down window we get back and add it to the var taskActionsEl.

    listItemEl.appendChild(taskActionsEl);
when then add this new taskActionsEl var to the listItemsEl <li> created
created above

    if (tasks[i].status === "to do") {
      listItemEl.querySelector("select[name='status-change']").selectedIndex = 0;
      tasksToDoEl.append(listItemEl);
    }
    else if (tasks[i].status === "in progress") {
      listItemEl.querySelector("select[name='status-change']").selectedIndex = 1;
      tasksInProgressEl.append(listItemEl);
    }
    else if (tasks[i].status === "completed") {
      listItemEl.querySelector("select[name='status-change']").selectedIndex = 2;
      tasksCompletedEl.appendChild(listItemEl);
    }
these all check the current arrays status to see that they are called.  Then it uses query selector to to go into it's <select> and sets it's windows text to the proper name <option> by using selectedIndex to assign it's value.  Example if "to do" was pulled from the saved array it will go into the <select> and choose <option> 0 which will then apply the text To Do to the drop down menu.
Finally it takes the entire li and adds it to the appropriate ul  in the html which are set up in the js globals.

    taskIdCounter ++;
then it increments the id number by 1.  This is important because say we reload 2 objects 1 of cat and one of dog in this array.  You would have an id of 0 for cat and an id1 for dog.   Then lets say in the current session we add a new <li> called bird, this will receive an id of 0.   This will function oddly because we will have to <li> with an id of zero.  By adding this it ensures that the global var taskIdCounter will first go up by one for everything that gets loaded.  Then in our current session whenever we create something new it will have the new higher id number.  In our example upon loading Dog would get assigned id 0, cat would get assigned id 1 and then anything new we add this session after this for loop, in this case bird will get assigned id 2.
    console.log(listItemEl);
  }
close for loop
};
close func